pyproject.toml
[project]
name = "crew-ai-marketing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai-tools>=0.60.0",
    "crewai[tools]>=0.141.0",
    "google-generativeai>=0.8.5",
    "langchain-openai>=0.2.14",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]

[tool.ruff]
ignore = ["E501"]
